# src/main/resources/api-specs/category-api-spec.yml
openapi: 3.1.0
info:
  title: Category API
  description: Category API
  version: 1.0.0

paths:
  /api/v1/categories:
    get:
      summary: Get all categories
      security:
        - BearerAuth: [ ]
      tags:
        - Category
      responses:
        200:
          description: List of all categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'

  /api/v1/categories/{id}:
    get:
      summary: Get category by ID
      security:
        - BearerAuth: [ ]
      tags:
        - Category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID of the category
      responses:
        200:
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        404:
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create Category
      security:
        - BearerAuth: [ ]
      description: Requires `ADMIN` role.
      tags:
        - Category
      requestBody:
        description: User data to create a new category
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'

      responses:
        200:
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'

        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    put:
      summary: Update Category
      security:
        - BearerAuth: [ ]
      description: Requires `ADMIN` role.
      tags:
        - Category
      requestBody:
        description: User data to update category
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'

      responses:
        200:
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    delete:
      summary: Delete a category
      security:
        - BearerAuth: [ ]
      description: Requires `ADMIN` role.
      tags:
        - Category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: long
          description: ID of the product
      responses:
        200:
          description: Category Deleted
        404:
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          description: URI reference identifying the error type
        title:
          type: string
          description: A short description of the error
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Detailed description of the error

    CategoryDto:
      type: object
      properties:
        title:
          type: string
          description: Title of the category (must include specific keywords and cannot exceed 50 characters)
          maxLength: 100
        description:
          type: string
          description: Detailed description of the category (cannot exceed 255 characters)
          maxLength: 255

    ValidationError:
      type: object
      properties:
        status:
          type: string
          example: '400'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ParamsViolationDetails'


    ParamsViolationDetails:
      type: object
      properties:
        fieldName:
          type: string
          description: Field where validation error occurred
        reason:
          type: string
          description: Reason for the validation error